refac: refactor custom model
