+++ tests/config/architectures/test_custom_model.py
@@ -1,22 +1,7 @@
 import pytest
-from torch import nn, ones
 
 from careamics.config.architectures import CustomModel, get_custom_model, register_model
 from careamics.config.support import SupportedArchitecture
-
-
-@register_model(name="linear")
-class LinearModel(nn.Module):
-    def __init__(self, in_features, out_features):
-        super().__init__()
-
-        self.in_features = in_features
-        self.out_features = out_features
-        self.weight = nn.Parameter(ones(in_features, out_features))
-        self.bias = nn.Parameter(ones(out_features))
-
-    def forward(self, input):
-        return (input @ self.weight) + self.bias
 
 
 @register_model(name="not_a_model")
@@ -37,9 +22,9 @@ def test_any_custom_parameters():
     CustomModel(architecture="custom", name="linear", in_features=10, out_features=5)
 
 
-def test_linear_model():
+def test_linear_model(custom_model_name):
     """Test that the model can be retrieved and instantiated."""
-    model = get_custom_model("linear")
+    model = get_custom_model(custom_model_name)
     model(in_features=10, out_features=5)
 
 
@@ -49,24 +34,15 @@ def test_not_a_model():
     model(3)
 
 
-def test_custom_model():
+def test_custom_model(custom_model_parameters):
     """Test that the custom model can be instantiated."""
-    # prepare model dictionary
-    model_dict = {
-        "architecture": SupportedArchitecture.CUSTOM.value,
-        "name": "linear",
-        "in_features": 10,
-        "out_features": 5,
-    }
-
     # create Pydantic model
-    pydantic_model = CustomModel(**model_dict)
+    pydantic_model = CustomModel(**custom_model_parameters)
 
     # instantiate model
     model_class = get_custom_model(pydantic_model.name)
     model = model_class(**pydantic_model.model_dump())
 
-    assert isinstance(model, LinearModel)
     assert model.in_features == 10
     assert model.out_features == 5
